pyproject.toml
[project]
name = "lifeos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.0.0",
    "groq>=1.0.0",
]
